feat(cwl): Add line trimming to LiveTail when limit is reached (#5837)
## Problem
* LiveTail sessions can run for a long time, on LogGroups that have a
high volume of LogEvents. We want to not infinitely add logs to a
TextDocument, leading to memory issues within VSCode.

## Solution
Users can configure a `limit` preference (default: 10000 , max 10000 ,
min: 1000).
When the number of lines in a Live Tail session (number of LogEvents)
reaches the limit, the 'N' oldest logEvents will be removed, to fit in
the 'N' newest events coming in from the response stream.