feat(amazonq): Add project context to inline completion (#5729)
## Problem

1. Inline Completion of Q is not context aware.
2. The `@workspace` index does not update on file add/remove.



## Solution

1. We are adding new global BM25 index and class/function digest from
imported files to provide project context for inline completion. This
feature will be added as AB testing. In this PR, there will be no
customer facing changes since it is hidden behind a AB testing flag. We
will toggle it a few days later. For this, we added new LSP APIs.

2. One customer facing change, when user add or remove a file in the VS
Code IDE, its `@workspace` index will be updated

The net-browserify is added to prevent npm run package failure for
browser.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Lei Gao <leigaol@amazon.com>
Co-authored-by: Lei Gao <97199248+leigaol@users.noreply.github.com>
Co-authored-by: Maxim Hayes <149123719+hayemaxi@users.noreply.github.com>