ci(lint): run jscpd in CI, compare to static threshold.  (#5833)
## Problem
New contributors often copy-paste logic and test cases, which creates a
maintenance cost for little gain.

Examples

-
https://github.com/aws/aws-toolkit-vscode/pull/3837#discussion_r1367182525
-
https://github.com/aws/aws-toolkit-vscode/pull/3805#discussion_r1317867006

We call these repeated segments of code "clones". 

## Solution
implement a static threshold. 
Run `jscpd` in CI and compare the % duplicated to a fixed threshold of
1.34%, which is the current duplication rate in the repo.
The main issue with this solution is that it fails to provide easy
debugging options when new duplicates are found. Currently exploring
adding this functionality in a follow-up.

#### Notes
- this is a summary of the problem/solution. For full information, see
full doc (VSCode: Detecting and Reducing Duplicate Code) describing
changes.
---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>