feat(amazonq): Add inline chat to context menu and reject empty inputs (#5879)
This PR adds the "Inline chat" menu option to the right click context
menu and command palette. It also rejects empty inputs to inline chat to
make it clearer to users they must input text into the dialog.


License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.