feat(notifications): setup and activation (#5932)
- Gated behind dev setting (`aws.dev.notifications`)
- Setup the notifications panel and begin polling in each of the
extensions.
- Currently the code collects a bunch of information about the state of
auth for telemetry (auth_userState). Refactor this code in each
extension so that we can re-use that information for the notification
rule engine.
- Change emitting auth_userState to node only for amazonq, since
notifications is node-only. Decided to punt this work for web because
spending effort to make it web-compatible may be in vain if the design
of web metrics is different.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.