telemetry(amazonq): Add more information for outlier chat round trips (#5944)
## Problem
It's hard to diagnose why we sometimes get some extreme values for a
chat round trip. One idea is that some remote instances might be hosted
in a different region and that's causing the increased values in some
rare instances

## Solution
Add the stringified version to the reasonDesc field temporarily so that
we can get more information. This would look like:
```
'{"chatMessageSent":1730994444010,"editorReceivedMessage":1730994444011,"featureReceivedMessage":1730994444013,"messageDisplayed":1730994459698}',
```

## Additional info
The numbers picked were pretty arbitrary `(x > 30000 || x < -30000)` but
it should give us a sampling of 50+ examples in the next week. After
that we can revert this code since we should have all the data we need
to see if this is the issue