build(lint): printf/console.log format specifiers must match args #5895
## Problem
currently `logger.error(msg, err)` silently drops `err` if it doesn't
map to a `%x` format specifier in msg.
we want to either prevent that mistake, or enhance our logger so that it
prints useful parts of err instead of silently dropping it.

Also, there are cases like:
`getLogger().error('Certificate path {0} already exists', certPath)`,
but this silently drops the arg `certPath`. Pretty sure this syntax is
from Java IIRC and not supported by Node.

## Solution
Provide a custom lint rule to ensure the number of extra args matches
the number of format specifiers.

### Note About Alternatives
- As an alternative, we could have transformed the extra `meta` items to
strings and appended them to the log message or add them as separate log
messages. This would allow usage of the form `getLogger().error('hit
error', e)`.
- If we add separate log messages, we break the expectation that each
`logger.error` statement corresponds to one message in the log.
- Conversely, if we append to the log message, this is somewhat of a
hidden feature not exposed to the consumer. By forcing them to add a
'%O' format specifier, we are making them clearly aware of how the error
(or other object) will show up in the log, ensuring the effect does not
"surprise" the user.

## Notes
Relevant: code used to format errors: 
https://github.com/aws/aws-toolkit-vscode/blob/f806ae04f5d56b872282e42d5f765b2b1a9e28dd/packages/core/src/shared/errors.ts#L329-L334
which should now actually be used since errors won't be dropped.