telemetry: add notifications telemetry (#5956)
I explored sending telemetry if the user hides or unhides the
notification view panel. However, there is no VSC hook to detect this
and other ways (e.g. checking visible property) are not reliable. So, we
cannot track this.

- toolkit_showNotification
  - for displaying notifications on receive or when clicked on
- toolkit_invokeAction
- invoking an action from a notification, e.g. clicking on a button in
the notification or opening a text document
- ui_click for clicking on the notifications node or dismiss button


---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.