fix(dev): dev menu not loading items #5972
## Problem
When opening the AWS Developer Menu for toolkit, no items were showing.

This is due to a wrong value being used for menu options. The second use
was expecting it to be an object with keys, when it was actually a
function. This happened due to a recent change that didn't catch this