fix(cwl): LiveTail fails to start when clicking with Play next to a LogGroup #5986
## Problem
Starting a LiveTail session by clicking Play next to a specific LogGroup
in the explorer menu is failing. This is because the LogGroup name is
passed into the Wizard response, and not a fully qualified Arn.

StartLiveTail API request requires ARNs.

## Solution
Detect if the context when initializing the TailLogGroup wizard is a
LogGroup Name or Arn. If it is just a name, construct the Arn and set
that in the Wizard response.

Renames `LogGroupName` in `LiveTailSession` to `LogGroupArn` to make it
more clear what is expected.