feat(amazonq): Enable implicit `@workspace` context for Amzn users #5856
## Problem
The chat workspace context requires user to manually input `@workspace`
in the chat window, but when user do not have such prompt, workspace
context can still be useful, we want to always send the relevant
document and let service decide when to utilize it.

## Solution

Enable implicit `@workspace` context for Amzn users. For these users, if
they do not have `@workspace` in the chat prompt, the
relevantTextDocument is still added to the chat API. This will be AB
tested

Also remove the outdated data collection experiment.