refactor(notifications): remove unneeded InstalledExtensions criteria
We have both ActiveExtensions (checks what extensions are installed and active), and InstalledExtensions (just what extensions are installed).
This removes InstalledExtensions because most of the intention behind it is captured by ActiveExtensions. When would we only care about whats installed but not active?

Removing a criteria decreases complexity and confusion in both code and for those drafting notifications, so this is an easy removal.