fix(cwl): pass credentials to LiveTail client #5993
## Problem
When creating the CloudWatchClient for a LiveTail session, we are not
specifying which credentials to use. This is causing the client to
always use the `Default` credential profile, even if a different AWS
Credential profile is selected within AWSToolkit.

## Solution
Resolve the active AWS credential from `globals.awsContext`, and supply
that to the LiveTailSession constructor. These credentials are then use
to construct the CWL client.