fix(appbuilder): misleading placeholder message #6010
## Problem
I was working on a CDK project, and after synthesizing my CDK template,
I noticed that the app-builder is showing a misleading notification to
customers: No IaC templates found in the workspace. This seems
incorrect, as I do have IaC templates in the workspace, including both
raw and synthesized CDK templates.

While I understand the intention behind the message, it needs to be more
precise to avoid confusion. I suggest we roll it back to something more
accurate, like No SAM templates found in the workspace, to clearly
indicate the specific missing template type.

## Solution
Solution is to change all instance where IAC is mentioned to the
customer to SAM, we would roll back to IaC once we start support CDK.

Remove redundant Logic.