fix(notifications): update types, field names, cross-window syncing (#6040)
- Make types more granular.
- Align names - fix onRecieve typo, openTxt -> openTextDocument
- Sync from global state on reload to detect activity in other windows.
This makes it so that if users dismiss in one window, it will dismiss in
all windows. It will also prevent a bug where users may get dismissed
notifications again if they have multiple windows.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.