fix(cwl): Hide LiveTail CodeLenses and display info message when closing session (#6063)
## Problem
Currently after closing a session in a way that leaves the TextDocument
open (`Stop tailing` CodeLens), the codeLenses remain visible. This
means a user could click `Stop tailing` again, and receive an error for
not being able to find the running LiveTail session for the given
document.

Additionally, there is no feedback when a session is closed. Clicking
`Stop tailing` doesn't signal to the user that the session was actually
stopped.

## Solution
* Provide a `refresh` method in the LiveTail CodeLens provider. This
fires an event to force recomputing the CodeLenses on a document.
* Modify LiveTail Lens provider to return no Lenses if the session is
not running (in the registry)
* Display an information window when a Session is stopped
* Changes wording/placement on some log statements for consistency.