feat(amazon q): removing feature flag for selective transformation and confirmed string changes (#6084)
## Problem
Selective transformation changes were released under an IDE feature flag
due to backend changes that needed to be fixed.

## Solution
In this PR, we are removing the feature flag on the IDE side so the
users will now be able to see the form to select if they want to view
their changes in one or multiple diffs. If they select multiple diffs in
this form, then selective transformation will be invoked.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.