feat(amazonq): support Dockerfile files in /dev (#6107)
## Problem
`Dockerfile` files are not currently supported by /dev

## Solution
Add a list of file prefixes that are considered code files for /dev.
This approach supports not only `Dockerfile` but also popular variants
of this such as `Dockerfile.build`, `Dockerfile-prod`, etc.

I manually tested some prompts in /dev with requests. I set breakpoints
to see that the file was being zipped up in workspace upload.