test: add notification and reset state dev options (#6105)
- Add a generic dev menu item to reset the state of a feature. Simply
extend with a function handler to add an item to the menu. The initial
item resets in-ide notification state.

![image](https://github.com/user-attachments/assets/042bdad2-2bac-4694-a7ff-801a857bddd3)

![image](https://github.com/user-attachments/assets/43fa90aa-af58-4b6f-af0d-ca38fb01f5b7)

- Add a way to test notifications locally without modifying code. Using
a dev menu item you can add a notification to global state and receive
it in your IDE for visualization.

![image](https://github.com/user-attachments/assets/2a0f267c-e792-4535-a1a4-170f8f6f32ca)

![image](https://github.com/user-attachments/assets/dd08402c-c4db-422c-9b60-20cd4f9340a2)

- Refactor notifications controller to not require a RuleEngine when
polling is requested.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.