test(cwl): Add unit test for starting and stopping LiveTailSession object.  (#6110)
## Problem
The `LiveTailSession` object's start/stop methods aren't directly under
test. TailLogGroup tests are mocking their behavior. TailLogGroup
command tests are more focused on managing the session registry,
opening/writing to the document, and the close tab behaviors.

## Solution
Write a test that's scoped to creating a `session` and calling start and
stop on it. Assert that timers are handled/disposed of properly, and
that the session response stream is returned.

Addresses [this
comment](https://github.com/aws/aws-toolkit-vscode/pull/6095#discussion_r1857571286).

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Keegan Irby <keegani@amazon.com>