telemetry(amazon q): metrics after one/multiple diffs form selection #6138
## Problem
We were only recording the metrics when the user accepts changes so we
can track which patches were applied based on their description. We
wanted to also keep track of the form in which the user is prompted to
pick whether they want their changes in one or multiple diffs.

## Solution
Adding a telemetry call when we handle the form selection for selective
transformation one or multiple diffs.