telemetry(cwl): LiveTail metrics
## Problem
Updating telemetry to accommodate changes in
https://github.com/aws/aws-toolkit-common/pull/932

## Solution
Update package version for telemetry. 
Add validation on the LogStreamFilter submenu's response for filter
type. This is needed to allow the type returned from the
LogStreamFilterSubmenu to be convertible to the type in the metric
definition for filterPattern. In any case, this is a good validation to
have since the 'menu' placeholder value isn't valid for starting a
LiveTail session anyways.